code is finished

+++ ex4/Ex41.c
@@ -5,10 +5,16 @@
 #include <stdlib.h>
 #include <stdbool.h>
 
+//macros
 #define READ 0
 #define WRITE 1
 #define EXIT_BUTTON 'q'
+#define RIGHT_BUTTON 'd'
+#define LEFT_BUTTON 'a'
+#define DOWN_BUTTON 's'
+#define SWAP_BUTTON 'w'
 
+//given function getch
 char getch(){
     char buf=0;
     struct termios old={0};
@@ -31,12 +37,13 @@ char getch(){
     return buf;
 }
 
+// exit gracfully
 void exitGracfully(int pipe[], int pid, int status){
     close(pipe[WRITE]);
     if (pid != 0){kill(pid, SIGKILL);}
     exit(status);
 }
-
+//handling fork error
 int handleForkError(int pipe[], int pid){
     if (pid < 0){ //check if error happened
         exitGracfully(pipe, 0, -1);
@@ -45,38 +52,36 @@ int handleForkError(int pipe[], int pid){
 }
 
 //function forks and runs draw proc
-//TODO handling errors
 int runGame(int pipe[]){
-    int childId = fork();
-    printf("%d", childId);
-    if (childId == 0){
+    int childId = fork(); //forking
+    if (childId == 0){ //child
         close(pipe[WRITE]);
         dup2(pipe[READ],0);
         close(pipe[READ]);
         execl("draw.out", "draw.out", NULL);
-    }
-    handleForkError(pipe, childId);
+    } 
+    handleForkError(pipe, childId); //father
     return childId;
-}
-void writeKey(char keyPressed){
-    printf("%c", keyPressed);
 }
 
 //handeling key presses
 void handleKeys(int pid, int* pipe){
     char keyPressed;
     while (true) {
-        keyPressed = getch();
-        write(pipe[WRITE], &keyPressed, 1);
-        if (keyPressed == EXIT_BUTTON) {exitGracfully(pipe, pid, 0); exit(-1);}
-        kill(pid, SIGUSR2);
+        keyPressed = getch(); //getting key press from user
+        //if it's a relevant key
+        if (keyPressed == RIGHT_BUTTON || keyPressed == LEFT_BUTTON || keyPressed == DOWN_BUTTON || keyPressed == SWAP_BUTTON || keyPressed == EXIT_BUTTON){
+            write(pipe[WRITE], &keyPressed, 1);
+            if (keyPressed == EXIT_BUTTON) {exitGracfully(pipe, pid, 0); exit(-1);}
+            kill(pid, SIGUSR2);
+        }
     }
 }
 
 int main() {
     int fd_pipe[2];
-    pipe(fd_pipe);
-    int childId = runGame(fd_pipe);
+    pipe(fd_pipe); //create pipe
+    int childId = runGame(fd_pipe); //start draw.out
     handleKeys(childId, fd_pipe);
     return 0;
 }
